Split AMDPARsing into a separate file
Also some minor cleanup
